Added TokenService class

+++ src/main/java/com/project/interfaces/ITeamService.java
@@ -12,5 +12,5 @@ public interface ITeamService {
 
     List<Team> getTeamsForUser(Integer id);
 
-    boolean deleteTeam(Long id);
+    boolean deleteTeam(Integer id);
 }
